==> analysis/1. STI_analysis.Rmd <==
---
title: "STI analysis"
output: html_notebook
---


```{r load library}
library(sp)
library(sf)
library(here)
library(dplyr)
library(ggplot2)
library(chron)
library(stringr)
library(ff)

```

In this chunk we create a grid for our bounding box of interest and we create our 0.2 grid

```{r create grid}

xmin <- -5.48
ymin <- 29.87
xmax <- 38.11
ymax <- 45.99

lonRange = c(-5.48,38.11)
latRange = c(29.87,45.99)

# set the desired grid cell size (in degrees)
cellsize <- 0.2

# create a bounding box polygon with fixed number of rows and columns
bbox <- st_polygon(list(rbind(c(xmin, ymin), c(xmin, ymax), c(xmax, ymax), c(xmax, ymin), c(xmin, ymin))))
polys <- st_make_grid(bbox, cellsize=cellsize)

# set the coordinate reference system to WGS84
st_crs(polys) <- st_crs("+proj=longlat +datum=WGS84")

# write the shapefile 
st_write(polys, here("shapefiles","med_grid_0.2.shp"))
```

Then we make a dataframe with the uniqe cells and their id. We call this dataframe **domain** and we add GSA to it 

```{r create df for grid}
grid.sar = read_sf(here("shapefiles", "med_grid_0.2.shp"))
gsas = read_sf(here('shapefiles', 'GSAs_simplified.shp'))

wgs = "+proj=longlat +datum=WGS84 +no_defs +ellps=WGS84 +towgs84=0,0,0"

st_crs(gsas) = wgs
st_crs(grid.sar) = wgs

grid.sar.fortify = fortify(grid.sar)

grid.sar.fortify$id = grid.sar$FID

gridll = st_coordinates(st_centroid(grid.sar)) %>% as.data.frame()

colnames(gridll) = c("lon", "lat")
df.coo = data.frame(FID = grid.sar.fortify$id, lon = gridll$lon, lat = gridll$lat)


df_points <- st_as_sf(df.coo, coords = c("lon", "lat"), crs = 4326)

# assign each centroid to a gsa
df_filtered <- st_intersection(df_points, gsas)
df_filtered$GSA<- as.numeric(str_sub(df_filtered$SECT_COD, - 2)       )

over.gsas<-df_filtered%>% dplyr::select(SMU_NAME,SECT_COD,FID)

final_with_grid_gsa = left_join(df.coo, over.gsas) 

# filter GSAs that are part of Black Sea
df.coo= final_with_grid_gsa %>%
  filter(!is.na(SECT_COD), SECT_COD!='GSA28',SECT_COD!='GSA29',SECT_COD!='GSA30') %>% dplyr::select(FID,lon,lat)

# create our final grid
domain = df.coo
colnames(domain)[1] = "id"

saveRDS(domain, here("data","domain_new.rData"))
```

Here we load the sar data and we aggregate it to our grid to perform STI 

```{r load sar normalised by overpasses}
data.sar=readRDS(here::here('data',"data_sar_norm.rData"))

data.sar$UTC =  as.numeric(chron(paste("01/", data.sar$month,"/", data.sar$year), 
                                 format = c(dates. = "D/M/Y")))
data.sar$year = as.numeric(as.character(years(data.sar$UTC)))
data.sar$month = as.numeric(months(data.sar$UTC))
data.sar$time = 12 * (data.sar$year - min(data.sar$year)) + data.sar$month

```



```{r aggregate sar data to our grid}

df_points <- st_as_sf(data.sar, coords = c("lon_bin", "lat_bin"), crs = 4326)

# assign each point to our grid and sum by id and time (month + year)
df_filtered <- st_join(df_points, grid.sar, join = st_intersects)

df_filtered$id = df_filtered$FID


sar.counts = data.frame(id = df_filtered$FID, 
                        time = data.sar$time,
                        n = data.sar$n_detections)
sar.counts = sar.counts[which(!is.na(sar.counts$id)),]
sar.counts = aggregate(data = sar.counts,
                            n ~ id + time, 
                       FUN = "sum")

grid.sar = left_join(sar.counts, 
                     data.frame(id = df.coo$FID, long = df.coo$lon, lat = df.coo$lat))

grid.sar = grid.sar[which(grid.sar$id %in% domain$id),]

saveRDS(grid.sar, here('data','grid.sar_new.rData'))
```

RUN STI

```{r run STI}
source(here::here("functions", "STI.R"))

domain = readRDS(here('data', "domain_new.rData"))
X = readRDS(here('data', "grid.sar_new.rData"))


# Objects
tt = sort(unique(X$time))
TT = length(tt)
ids = domain$id
n = length(ids)
#
xj = rep(0, n*TT)
counter = 0
for(iTime in tt){
 for(cellID in ids){
  counter = counter + 1
  indice = which((X$id == cellID) & (X$time == iTime))
  if(length(indice) == 1) xj[counter] = X[indice,which(colnames(X)=='n')]
 }
}

# Space–time connectivity matrix
Ws = GetW(coords=domain, id.name='id', buffer=5, include.self=star)
VST = STCM(spatCon=Ws, nCells=n, nTimes=TT, include.self=star)

# Values of the index
Gj = Gstandard(stConn=VST, values=xj, nCells=n, include.self=star)
#create df of sti 
temp = cbind(domain, Gj)

pos_temp = temp %>% dplyr::filter(Gj>0)

hist(temp$Gj)
hist(pos_temp$Gj)

ggplot(temp, aes(x=Gj)) +
    geom_histogram(binwidth=1, colour="black", fill="white") +
    geom_vline(aes(xintercept=mean(Gj, na.rm=T)),   # Ignore NA values for mean
               color="red", linetype="dashed", size=1)

```

Plot STI results 

```{r}
library(viridis)
library(scales)
library(rnaturalearth)

estados <- ne_countries(type = "countries", scale = "small", returnclass = 'sf')


temp = readRDS(here('data/STI', 'temp.RData'))


temp$lat<- as.numeric(as.character(temp$lat))

temp$lon<- as.numeric(as.character(temp$lon))


xmin <- -5.48
ymin <- 29.87
xmax <- 38.11
ymax <- 45.99


p<-temp %>% 
  ggplot() +
  geom_tile(aes(x = lon, y = lat, fill = Gj), show.legend = TRUE) +
  scale_fill_viridis(limits = c(1, 40), breaks = seq(1, 40, by = 5), oob = scales::squish, na.value = NA) +
  geom_sf(data = estados, lwd = 0,fill='#F6F6F4', color='#F6F6F4') +
  
  
  theme(panel.background = element_rect(fill = "#D4F1F4"),
        panel.grid.major = element_line(color = "#B2BEB5"),
      plot.title = element_text(family = 'Roboto',
                                face = 'bold',
                                color = '#363c4c',
                                size = 20),
      plot.subtitle = element_text(family = 'Roboto',
                                   color = '#363c4c',
                                   size = 18),

      strip.text = element_text(family = 'Roboto',
                                face = 'bold',
                                color = '#363c4c',
                                size = 14),
        
      legend.text = element_text(family = 'Roboto',
                                 color = '#848b9b',
                                 size = 14),
      legend.title = element_text(family = 'Roboto',
                                  face = 'bold',
                                  color = '#363c4c',
                                  size = 16),
      legend.position = 'bottom',
      legend.box = 'vertical',
      legend.key.height = unit(3, 'mm'),
      legend.key.width = unit(20,'mm'),
      axis.title = element_text(family = 'Roboto',
                                face = 'bold',
                                color = '#848b9b',
                                size = 16),
      axis.text = element_text(family = 'Roboto',
                               color = '#848b9b',
                               size = 14),
      strip.background = element_rect(color="#d6c7db", fill="#d6c7db", linetype="solid"
     ),
     plot.margin=grid::unit(c(0,0,0,0), "null")) + 
  labs(x = "Longitude", y= "Latitude",fill='G* Index') +
  coord_sf(xlim = lonRange, ylim = latRange)   

p

ggsave(filename = here::here('figures', "index_all_med_0.25_all_years.png"),width = 30, height = 20, units = "cm")

```
